Refactor Results component.
Repetitive code in the Results component was consolidated, and certain
functionality was moved into separate components for modularity and
re-use.

The ResultsTemplate component formats the header and body of Results consistently so
that Results reflecting all users and individual users are easily
created. The Results component now uses this for loading the head and
body.

The IndividualResults component houses all the FilmCard components
representing an individual user's answers. It makes use of
ResultsTemplate to have a consistent look with the Results Component.

loadHead() and loadBody() were refactored into one function, loadBody(). Similar conditionals were in each, and through the ResultsTemplate and IndividualResults components the Results component was simplified.